improvement regurding shifting logic to hook and devided ui into smaller components
